Move getConflictsByTestResults
